refactor: Change NotFoundUserException package structure
