+++ src/main/java/com/practice/boardproject/member/error/NotFoundUserException.java
@@ -1,4 +1,4 @@
-package com.practice.boardproject.post.domain.exception;
+package com.practice.boardproject.member.error;
 
 public class NotFoundUserException extends RuntimeException {
     private static final String ERROR_MESSAGE_FORMAT = "해당 사용자를 찾을 수 없습니다: %s";
